requirements.txt
alembic==1.0.7
aniso8601==4.1.0
antlr4-python3-runtime==4.7.2
arrow==0.13.1
attrs==18.2.0
bcrypt==3.1.6
bencode.py==2.1.0
bencoder==0.2.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
fastjsonschema==2.11
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==3.2.4
Flask-JWT-Extended==3.17.0
Flask-Migrate==2.4.0
Flask-PyMongo==2.2.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
furl==2.0.0
future==0.17.1
geoip2==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0
lomond==0.3.3
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.19.2
maxminddb==1.4.1
medallion==0.3.1
numpy==1.17.2
orderedmultidict==1.0.1
pandas==0.25.1
passlib==1.7.1
pycparser==2.19
pycryptodome==3.8.0
PyJWT==1.7.1
pymisp==2.4.103
pymongo==3.8.0
pyperclip==1.7.0
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
python-editor==1.0.4
python-whois==0.7.2
pytz==2018.9
requests==2.22.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.2.18
stix2==1.1.2
stix2-patterns==1.1.0
tahoe==0.3
taxii2-client==0.5.0
urllib3==1.24.1
websocket-client==0.54.0
Werkzeug==0.14.1
